api/requirements.txt
# Core dependencies
flask==2.2.3
werkzeug==2.2.3
flask-sqlalchemy==3.0.3
flask-cors==3.0.10

# Development tools
python-dotenv==1.0.0

# Data processing (basic)
numpy==1.24.2
pandas==1.5.3
